% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preservelabels.R
\name{preservelabels}
\alias{preservelabels}
\title{Preserve the variable labels}
\usage{
preservelabels(data)
}
\arguments{
\item{data}{The data you want to preserve its variable labels}
}
\value{
list
}
\description{
Preserve the variable labels of a given dataset
before conducting manipulation that drops those
the labels
}
\examples{
preservelabels(labelspreserve::sample_data)
}
