% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restorelabels.R
\name{restorelabels}
\alias{restorelabels}
\title{Restore the variable labels}
\usage{
restorelabels(data, variable_labels = data_labels)
}
\arguments{
\item{data}{The data you want to restore its variable labels}

\item{variable_labels}{The variable labels saved in a list}
}
\value{
data.frame
}
\description{
Restore the variable labels
}
\examples{
preservelabels(labelspreserve::sample_data)
restorelabels(labelspreserve::sample_data)
}
