% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sampled data.}
\format{
A data frame with 200 rows and 8 variables:
\describe{
\item{trt}{Chemotherapy Treatment}
\item{age}{Age}
\item{marker}{Marker Level}
\item{stage}{T Stage}
\item{grade}{Grade}
\item{response}{Tumor Response}
\item{death}{Patient Died}
\item{ttdeath}{Months to Death/Censor}
...
}
}
\usage{
sample_data
}
\description{
A dataset on trial.
}
\keyword{datasets}
